Remove old PD (Prefill-Decode) routing mode, consolidate on vLLM PD mode
This change removes the old HTTP/gRPC-based PD routing implementation
and consolidates exclusively on the vLLM PD mode which uses ZMQ-based
KV transfer with service discovery.

Changes:
- Remove RoutingMode::PrefillDecode enum variant
- Delete src/routers/http/pd_router.rs (old HTTP PD router)
- Delete src/routers/grpc/pd_router.rs (old gRPC PD router)
- Remove --pd-disaggregation CLI flag from main.rs and lib.rs
- Update factory.rs to remove create_pd_router methods
- Update validation.rs to remove PrefillDecode validation
- Update module declarations (http/mod.rs, grpc/mod.rs, routers/mod.rs)
- Update service_discovery.rs to use VllmPDRouter instead of PDRouter
- Update server.rs to use create_vllm_pd_router
- Make vllm_pd_router.rs standalone (no longer depends on PDRouter)
- Fix unused code warnings (prefix unused items with underscore)

The vLLM PD mode (--vllm-pd-disaggregation) remains fully functional
and is now the only supported PD disaggregation mode.
