+++ src/config/types.rs
@@ -120,19 +120,6 @@ pub enum RoutingMode {
         /// List of worker URLs
         worker_urls: Vec<String>,
     },
-    #[serde(rename = "prefill_decode")]
-    PrefillDecode {
-        /// Prefill worker URLs with optional bootstrap ports
-        prefill_urls: Vec<(String, Option<u16>)>,
-        /// Decode worker URLs
-        decode_urls: Vec<String>,
-        /// Optional separate policy for prefill workers
-        #[serde(skip_serializing_if = "Option::is_none")]
-        prefill_policy: Option<PolicyConfig>,
-        /// Optional separate policy for decode workers
-        #[serde(skip_serializing_if = "Option::is_none")]
-        decode_policy: Option<PolicyConfig>,
-    },
     #[serde(rename = "openai")]
     OpenAI {
         /// OpenAI-compatible API base(s), provided via worker URLs
@@ -158,10 +145,7 @@ pub enum RoutingMode {
 
 impl RoutingMode {
     pub fn is_pd_mode(&self) -> bool {
-        matches!(
-            self,
-            RoutingMode::PrefillDecode { .. } | RoutingMode::VllmPrefillDecode { .. }
-        )
+        matches!(self, RoutingMode::VllmPrefillDecode { .. })
     }
 
     pub fn is_vllm_pd_mode(&self) -> bool {
@@ -171,11 +155,6 @@ impl RoutingMode {
     pub fn worker_count(&self) -> usize {
         match self {
             RoutingMode::Regular { worker_urls } => worker_urls.len(),
-            RoutingMode::PrefillDecode {
-                prefill_urls,
-                decode_urls,
-                ..
-            } => prefill_urls.len() + decode_urls.len(),
             RoutingMode::VllmPrefillDecode {
                 prefill_urls,
                 decode_urls,
@@ -190,9 +169,6 @@ impl RoutingMode {
     /// Falls back to the main policy if no specific prefill policy is set
     pub fn get_prefill_policy<'a>(&'a self, main_policy: &'a PolicyConfig) -> &'a PolicyConfig {
         match self {
-            RoutingMode::PrefillDecode { prefill_policy, .. } => {
-                prefill_policy.as_ref().unwrap_or(main_policy)
-            }
             RoutingMode::VllmPrefillDecode { prefill_policy, .. } => {
                 prefill_policy.as_ref().unwrap_or(main_policy)
             }
@@ -204,9 +180,6 @@ impl RoutingMode {
     /// Falls back to the main policy if no specific decode policy is set
     pub fn get_decode_policy<'a>(&'a self, main_policy: &'a PolicyConfig) -> &'a PolicyConfig {
         match self {
-            RoutingMode::PrefillDecode { decode_policy, .. } => {
-                decode_policy.as_ref().unwrap_or(main_policy)
-            }
             RoutingMode::VllmPrefillDecode { decode_policy, .. } => {
                 decode_policy.as_ref().unwrap_or(main_policy)
             }
@@ -462,7 +435,6 @@ impl RouterConfig {
     pub fn mode_type(&self) -> &'static str {
         match self.mode {
             RoutingMode::Regular { .. } => "regular",
-            RoutingMode::PrefillDecode { .. } => "prefill_decode",
             RoutingMode::VllmPrefillDecode { .. } => "vllm_prefill_decode",
             RoutingMode::OpenAI { .. } => "openai",
         }
@@ -588,11 +560,12 @@ mod tests {
         };
         assert!(!regular.is_pd_mode());
 
-        let pd = RoutingMode::PrefillDecode {
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![("http://prefill1".to_string(), Some(8001))],
             decode_urls: vec!["http://decode1".to_string()],
             prefill_policy: None,
             decode_policy: None,
+            discovery_address: None,
         };
         assert!(pd.is_pd_mode());
     }
@@ -608,7 +581,7 @@ mod tests {
         };
         assert_eq!(regular.worker_count(), 3);
 
-        let pd = RoutingMode::PrefillDecode {
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![
                 ("http://prefill1".to_string(), Some(8001)),
                 ("http://prefill2".to_string(), None),
@@ -620,6 +593,7 @@ mod tests {
             ],
             prefill_policy: None,
             decode_policy: None,
+            discovery_address: None,
         };
         assert_eq!(pd.worker_count(), 5);
 
@@ -639,15 +613,16 @@ mod tests {
         assert!(json.contains("\"type\":\"regular\""));
         assert!(json.contains("\"worker_urls\""));
 
-        // Test PrefillDecode mode
-        let pd = RoutingMode::PrefillDecode {
+        // Test VllmPrefillDecode mode
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![("http://prefill1".to_string(), Some(8001))],
             decode_urls: vec!["http://decode1".to_string()],
             prefill_policy: None,
             decode_policy: None,
+            discovery_address: None,
         };
         let json = serde_json::to_string(&pd).unwrap();
-        assert!(json.contains("\"type\":\"prefill_decode\""));
+        assert!(json.contains("\"type\":\"vllm_prefill_decode\""));
         assert!(json.contains("\"prefill_urls\""));
         assert!(json.contains("\"decode_urls\""));
     }
@@ -838,15 +813,16 @@ mod tests {
         assert_eq!(config.mode_type(), "regular");
 
         let config = RouterConfig {
-            mode: RoutingMode::PrefillDecode {
+            mode: RoutingMode::VllmPrefillDecode {
                 prefill_urls: vec![],
                 decode_urls: vec![],
                 prefill_policy: None,
                 decode_policy: None,
+                discovery_address: None,
             },
             ..Default::default()
         };
-        assert_eq!(config.mode_type(), "prefill_decode");
+        assert_eq!(config.mode_type(), "vllm_prefill_decode");
     }
 
     #[test]
@@ -955,9 +931,9 @@ mod tests {
     // ============= Complex Configuration Tests =============
 
     #[test]
-    fn test_full_pd_mode_config() {
+    fn test_full_vllm_pd_mode_config() {
         let config = RouterConfig {
-            mode: RoutingMode::PrefillDecode {
+            mode: RoutingMode::VllmPrefillDecode {
                 prefill_urls: vec![
                     ("http://prefill1:8000".to_string(), Some(8001)),
                     ("http://prefill2:8000".to_string(), None),
@@ -968,6 +944,7 @@ mod tests {
                 ],
                 prefill_policy: None,
                 decode_policy: None,
+                discovery_address: None,
             },
             policy: PolicyConfig::PowerOfTwo {
                 load_check_interval_secs: 30,
@@ -1158,9 +1135,9 @@ mod tests {
     // ============= Policy Fallback Tests =============
 
     #[test]
-    fn test_pd_policy_fallback_both_specified() {
+    fn test_vllm_pd_policy_fallback_both_specified() {
         // When both prefill and decode policies are specified, they should be used
-        let pd = RoutingMode::PrefillDecode {
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![("http://prefill1".to_string(), None)],
             decode_urls: vec!["http://decode1".to_string()],
             prefill_policy: Some(PolicyConfig::CacheAware {
@@ -1173,6 +1150,7 @@ mod tests {
             decode_policy: Some(PolicyConfig::PowerOfTwo {
                 load_check_interval_secs: 60,
             }),
+            discovery_address: None,
         };
 
         let main_policy = PolicyConfig::Random;
@@ -1190,9 +1168,9 @@ mod tests {
     }
 
     #[test]
-    fn test_pd_policy_fallback_only_prefill() {
+    fn test_vllm_pd_policy_fallback_only_prefill() {
         // When only prefill policy is specified, decode should use main policy
-        let pd = RoutingMode::PrefillDecode {
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![("http://prefill1".to_string(), None)],
             decode_urls: vec!["http://decode1".to_string()],
             prefill_policy: Some(PolicyConfig::CacheAware {
@@ -1203,6 +1181,7 @@ mod tests {
                 max_tree_size: 1000,
             }),
             decode_policy: None,
+            discovery_address: None,
         };
 
         let main_policy = PolicyConfig::RoundRobin;
@@ -1221,15 +1200,16 @@ mod tests {
     }
 
     #[test]
-    fn test_pd_policy_fallback_only_decode() {
+    fn test_vllm_pd_policy_fallback_only_decode() {
         // When only decode policy is specified, prefill should use main policy
-        let pd = RoutingMode::PrefillDecode {
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![("http://prefill1".to_string(), None)],
             decode_urls: vec!["http://decode1".to_string()],
             prefill_policy: None,
             decode_policy: Some(PolicyConfig::PowerOfTwo {
                 load_check_interval_secs: 60,
             }),
+            discovery_address: None,
         };
 
         let main_policy = PolicyConfig::Random;
@@ -1248,13 +1228,14 @@ mod tests {
     }
 
     #[test]
-    fn test_pd_policy_fallback_none_specified() {
+    fn test_vllm_pd_policy_fallback_none_specified() {
         // When no specific policies are specified, both should use main policy
-        let pd = RoutingMode::PrefillDecode {
+        let pd = RoutingMode::VllmPrefillDecode {
             prefill_urls: vec![("http://prefill1".to_string(), None)],
             decode_urls: vec!["http://decode1".to_string()],
             prefill_policy: None,
             decode_policy: None,
+            discovery_address: None,
         };
 
         let main_policy = PolicyConfig::CacheAware {
