+++ src/routers/mod.rs
@@ -22,7 +22,7 @@ pub mod router_manager;
 
 pub use factory::RouterFactory;
 // Re-export HTTP routers for convenience (keeps routers::openai_router path working)
-pub use http::{openai_router, pd_router, pd_types, router};
+pub use http::{openai_router, pd_types, router};
 
 /// Worker management trait for administrative operations
 ///
